Part 6

+++ app/controllers/questions_controller.rb
@@ -1,19 +1,23 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+  
   def index
     @questions = Question.all
   end
 
   def show
+    @answer = question.answers.new
   end
 
   def new
+    @question = Question.new
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.build(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Question created'
     else
       render :new
     end
@@ -31,8 +35,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question deleted'
+    else
+      redirect_to question, alert: 'Question cannot be deleted'
+    end
   end
 
   private
